# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quiniela Mundial 2026
  description: |
    API REST para la aplicación de Quiniela del Mundial Norteamérica 2026.

    **Reglas de negocio clave:**
    - Las predicciones se cierran 15 minutos antes del `start_time` del partido (estado `Locked`).
    - Las predicciones de otros usuarios no son visibles hasta que el partido esté `Locked` o `Finished`.
    - Puntuación: +3 por tendencia (1X2) correcta, +2 de bono por marcador exacto (total 5 pts).
    - Las bonus questions se bloquean al inicio del partido inaugural del torneo.

    **Time Travel:** Todos los endpoints que dependen del tiempo aceptan `simulated_time` (ISO 8601)
    para facilitar pruebas E2E. Requiere `ENABLE_TIME_TRAVEL=True` en el backend.
  version: 1.0.0
  contact:
    name: Quiniela Dev Team
servers:
  - url: http://localhost:8000
    description: Desarrollo local
...
paths:
...
  /users/me:
    get:
      summary: Perfil del usuario autenticado
      operationId: getMe
      tags: [users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Datos del usuario con su posición en el ranking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMe'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/ranking:
    get:
      summary: Tabla de posiciones global
      operationId: getRanking
      tags: [users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista ordenada por `total_points DESC, exact_scores DESC, username ASC`
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RankingEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Matches ───────────────────────────────────────────────────────────────
  /matches:
    get:
      summary: Listar partidos
      operationId: listMatches
      tags: [matches]
      security:
        - bearerAuth: []
      parameters:
        - name: phase
          in: query
          description: "Filtrar por fase: Groups | R32 | R16 | QF | SF | ThirdPlace | Final"
          schema:
            type: string
            enum: [Groups, R32, R16, QF, SF, ThirdPlace, Final]
        - name: group_name
          in: query
          description: "Filtrar por grupo (A–L, solo aplica en phase=Groups)"
          schema:
            type: string
        - name: simulated_time
          in: query
          description: "Time Travel — ISO 8601 (requiere ENABLE_TIME_TRAVEL=True en backend)"
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Lista de partidos ordenada por `start_time ASC`. El campo `status` es calculado en runtime.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchOut'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /matches/{match_id}:
    get:
      summary: Detalle de un partido
      operationId: getMatch
      tags: [matches]
      security:
        - bearerAuth: []
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: integer
        - name: simulated_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Partido con predicción del usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchOut'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Predictions ───────────────────────────────────────────────────────────
  /predictions:
    post:
      summary: Crear o actualizar predicción (upsert)
      description: |
        Crea una predicción nueva o actualiza la existente para el par `user_id + match_id`.
        Solo disponible cuando el partido está en estado `Open` (más de 15 min antes del pitazo).
      operationId: upsertPrediction
      tags: [predictions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionCreate'
      responses:
        '200':
          description: Predicción creada o actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionOut'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Partido cerrado (Locked o Finished)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example:
                detail: Match is locked. Predictions closed.
        '404':
          $ref: '#/components/responses/NotFound'
  /predictions/match/{match_id}:
    get:
      summary: Ver predicciones de todos los usuarios para un partido
      description: |
        Solo accesible cuando el partido está en estado `Locked` o `Finished`.
        Protege la privacidad de los jugadores hasta que el partido comienza.
      operationId: getMatchPredictions
      tags: [predictions]
      security:
        - bearerAuth: []
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: integer
        - name: simulated_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Lista de predicciones de todos los usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendPrediction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Partido aún abierto — predicciones no visibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example:
                detail: Predictions are only visible after the match is locked.
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Bonus ─────────────────────────────────────────────────────────────────
...
  /admin/matches:
    post:
      summary: Crear partido
      operationId: adminCreateMatch
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchCreate'
      responses:
        '201':
          description: Partido creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchOut'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/matches/{match_id}:
    put:
      summary: Actualizar datos del partido (fecha, equipos, sede)
      operationId: adminUpdateMatch
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchUpdate'
      responses:
        '200':
          description: Partido actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchOut'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: Could not validate credentials
    Forbidden:
      description: El usuario no tiene permisos de administrador
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: Admin access required
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
  schemas:
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
      required: [detail]

    # ── Auth schemas ──────────────────────────────────────────────────────────
...
    UserMe:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        is_admin:
          type: boolean
        status:
          type: string
          enum: [Pending, Active, Rejected]
        total_points:
          type: integer
        exact_scores:
          type: integer
        rank:
          type: integer
          nullable: true
    RankingEntry:
      type: object
      properties:
        rank:
          type: integer
        user_id:
          type: integer
        username:
          type: string
        total_points:
          type: integer
        exact_scores:
          type: integer
...
    MatchOut:
      type: object
      properties:
        id:
          type: integer
        home_team:
          type: string
          nullable: true
          description: Nulo en fases eliminatorias hasta que se definan los equipos
        away_team:
          type: string
          nullable: true
        home_team_code:
          type: string
          nullable: true
          description: "ISO 3166-1 alpha-2 para construir URL de bandera (flagcdn.com)"
        away_team_code:
          type: string
          nullable: true
        home_team_placeholder:
          type: string
          nullable: true
          description: "Texto descriptivo cuando home_team es null, ej: '1° Grupo A'"
        away_team_placeholder:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        phase:
          type: string
          enum: [Groups, R32, R16, QF, SF, ThirdPlace, Final]
        group_name:
          type: string
          nullable: true
          description: Solo aplica en phase=Groups (A–L)
        matchday:
          type: integer
          nullable: true
          description: Jornada dentro del grupo (1–3)
        venue:
          type: string
          nullable: true
        status:
          type: string
          enum: [Open, Locked, Finished, Pending Teams]
          description: |
            Estado calculado en runtime (no almacenado en DB):
            - Open: más de 15 min antes del pitazo
            - Locked: 15 min antes del pitazo hasta el final del partido
            - Finished: resultado cargado por Admin
            - Pending Teams: equipos eliminatorios aún no definidos
        home_score_final:
          type: integer
          nullable: true
        away_score_final:
          type: integer
          nullable: true
        my_prediction:
          $ref: '#/components/schemas/PredictionInMatch'
          nullable: true
    PredictionInMatch:
      type: object
      properties:
        home_score_guess:
          type: integer
        away_score_guess:
          type: integer
        points_earned:
          type: integer
          nullable: true
    MatchCreate:
      type: object
      required: [start_time, phase]
      properties:
        home_team:
          type: string
          nullable: true
        away_team:
          type: string
          nullable: true
        home_team_code:
          type: string
          nullable: true
        away_team_code:
          type: string
          nullable: true
        home_team_placeholder:
          type: string
          nullable: true
        away_team_placeholder:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        phase:
          type: string
          enum: [Groups, R32, R16, QF, SF, ThirdPlace, Final]
        group_name:
          type: string
          nullable: true
        matchday:
          type: integer
          nullable: true
        venue:
          type: string
          nullable: true
    MatchUpdate:
      type: object
      description: Todos los campos son opcionales — se aplica un patch parcial
      properties:
        home_team:
          type: string
          nullable: true
        away_team:
          type: string
          nullable: true
        home_team_code:
          type: string
          nullable: true
        away_team_code:
          type: string
          nullable: true
        home_team_placeholder:
          type: string
          nullable: true
        away_team_placeholder:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        venue:
          type: string
          nullable: true
...
    PredictionCreate:
      type: object
      required: [match_id, home_score_guess, away_score_guess]
      properties:
        match_id:
          type: integer
        home_score_guess:
          type: integer
          minimum: 0
        away_score_guess:
          type: integer
          minimum: 0
        simulated_time:
          type: string
          format: date-time
          nullable: true
          description: Time Travel — ISO 8601 (requiere ENABLE_TIME_TRAVEL=True)
    PredictionOut:
      type: object
      properties:
        id:
          type: integer
        match_id:
          type: integer
        home_score_guess:
          type: integer
        away_score_guess:
          type: integer
        points_earned:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FriendPrediction:
      type: object
      properties:
        username:
          type: string
        home_score_guess:
          type: integer
        away_score_guess:
          type: integer
        points_earned:
          type: integer
          nullable: true

    # ── Bonus schemas ─────────────────────────────────────────────────────────
...